zod validation with code refactor 🔥
