+++ logs/winston/success/.8631c91f803e76c37c77024a2d58e0af0acb3b99-audit.json
@@ -1,20 +1,20 @@
 {
-    "keep": {
-        "days": true,
-        "amount": 14
+  "keep": {
+    "days": true,
+    "amount": 14
+  },
+  "auditLog": "/workspaces/University-Management/logs/winston/success/.8631c91f803e76c37c77024a2d58e0af0acb3b99-audit.json",
+  "files": [
+    {
+      "date": 1686545043479,
+      "name": "/workspaces/University-Management/logs/winston/success/success-2023-12-06-04.log",
+      "hash": "781d5c0ce05092028e84087e3cc791ff77d993582caec31475fe3df3fa268be8"
     },
-    "auditLog": "/workspaces/University-Management/logs/winston/success/.8631c91f803e76c37c77024a2d58e0af0acb3b99-audit.json",
-    "files": [
-        {
-            "date": 1686545043479,
-            "name": "/workspaces/University-Management/logs/winston/success/success-2023-12-06-04.log",
-            "hash": "781d5c0ce05092028e84087e3cc791ff77d993582caec31475fe3df3fa268be8"
-        },
-        {
-            "date": 1686546008476,
-            "name": "/workspaces/University-Management/logs/winston/success/success-2023-12-06-05.log",
-            "hash": "394f654655b870daddc2f2b53dbc720ddde2ea1279114512df0c81694b035b73"
-        }
-    ],
-    "hashType": "sha256"
+    {
+      "date": 1686546008476,
+      "name": "/workspaces/University-Management/logs/winston/success/success-2023-12-06-05.log",
+      "hash": "394f654655b870daddc2f2b53dbc720ddde2ea1279114512df0c81694b035b73"
+    }
+  ],
+  "hashType": "sha256"
 }
