+++ package.json
@@ -36,6 +36,7 @@
     "express": "^4.18.2",
     "mongoose": "^7.2.1",
     "winston": "^3.9.0",
-    "winston-daily-rotate-file": "^4.7.1"
+    "winston-daily-rotate-file": "^4.7.1",
+    "zod": "^3.21.4"
   }
 }
