+++ src/app.ts
@@ -1,6 +1,6 @@
 import express, { Application } from 'express'
 import cors from 'cors'
-import { UserRoutes } from './app/modules/users/users.route'
+import { UserRoutes } from './app/modules/user/user.route'
 import globalErrorHandler from './app/middleware/globalErrorHandler'
 
 const app: Application = express()
