+++ src/app/modules/user/user.route.ts
@@ -1,7 +1,7 @@
 // route
 
 import { Router } from 'express'
-import { UserController } from './users.controller'
+import { UserController } from './user.controller'
 
 const router = Router()
